fix small issues
git-svn-id: https://osmand.googlecode.com/svn/trunk@318 e29c36b1-1cfa-d876-8d93-3434fc2bb7b8

+++ OsmAnd/src/com/osmand/views/OsmandMapTileView.java
@@ -315,7 +315,7 @@ public class OsmandMapTileView extends SurfaceView implements IMapDownloaderCall
 	//////////////////////////////// DRAWING MAP PART /////////////////////////////////////////////
 	
 	protected void drawEmptyTile(Canvas cvs, float x, float y){
-		int tileDiv = (int) (getTileSize() / emptyTileDivisor);
+		float tileDiv = (getTileSize() / emptyTileDivisor);
 		for (int k1 = 0; k1 < emptyTileDivisor; k1++) {
 			for (int k2 = 0; k2 < emptyTileDivisor; k2++) {
 				float xk = x + tileDiv* k1 ;
